chg: moved mappers to base package
