+++ src/main/java/sk/leziemevpezinku/spring/mapper/EventConditionMapper.java
@@ -1,9 +1,9 @@
-package sk.leziemevpezinku.spring.api.mapper;
+package sk.leziemevpezinku.spring.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingConstants;
 import sk.leziemevpezinku.spring.api.dto.EventConditionDTO;
-import sk.leziemevpezinku.spring.api.mapper.base.BaseMapper;
+import sk.leziemevpezinku.spring.mapper.base.BaseMapper;
 import sk.leziemevpezinku.spring.model.EventCondition;
 
 @Mapper(
