+++ src/main/java/sk/leziemevpezinku/spring/mapper/PeriodMapper.java
@@ -1,9 +1,9 @@
-package sk.leziemevpezinku.spring.api.mapper;
+package sk.leziemevpezinku.spring.mapper;
 
 import org.mapstruct.*;
 import sk.leziemevpezinku.spring.api.dto.PeriodDTO;
-import sk.leziemevpezinku.spring.api.mapper.annotation.IgnoreAuditedProperties;
-import sk.leziemevpezinku.spring.api.mapper.base.BaseMapper;
+import sk.leziemevpezinku.spring.mapper.annotation.IgnoreAuditedProperties;
+import sk.leziemevpezinku.spring.mapper.base.BaseMapper;
 import sk.leziemevpezinku.spring.model.Period;
 
 @Mapper(
