+++ src/main/java/sk/leziemevpezinku/spring/service/impl/EnumerationServiceBean.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sk.leziemevpezinku.spring.api.dto.EnumerationItemDTO;
 import sk.leziemevpezinku.spring.api.dto.PlaceDTO;
-import sk.leziemevpezinku.spring.api.mapper.EnumerationItemMapper;
-import sk.leziemevpezinku.spring.api.mapper.PlaceMapper;
+import sk.leziemevpezinku.spring.mapper.EnumerationItemMapper;
+import sk.leziemevpezinku.spring.mapper.PlaceMapper;
 import sk.leziemevpezinku.spring.model.EnumerationItem;
 import sk.leziemevpezinku.spring.api.enumeration.EnumerationName;
 import sk.leziemevpezinku.spring.repo.EnumerationRepository;
