+++ src/main/java/sk/leziemevpezinku/spring/service/impl/EventTermServiceBean.java
@@ -3,7 +3,7 @@ package sk.leziemevpezinku.spring.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import sk.leziemevpezinku.spring.api.dto.EventTermDTO;
-import sk.leziemevpezinku.spring.api.mapper.EventTermMapper;
+import sk.leziemevpezinku.spring.mapper.EventTermMapper;
 import sk.leziemevpezinku.spring.repo.EventRepository;
 import sk.leziemevpezinku.spring.repo.EventTermRepository;
 import sk.leziemevpezinku.spring.service.EventTermService;
