+++ src/main/java/sk/leziemevpezinku/spring/service/impl/PeriodServiceBean.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import sk.leziemevpezinku.spring.api.dto.PeriodDTO;
-import sk.leziemevpezinku.spring.api.mapper.PeriodMapper;
+import sk.leziemevpezinku.spring.mapper.PeriodMapper;
 import sk.leziemevpezinku.spring.model.Period;
 import sk.leziemevpezinku.spring.repo.PeriodRepository;
 import sk.leziemevpezinku.spring.service.PeriodService;
