+++ src/test/java/sk/leziemevpezinku/spring/api/EnumerationItemMapperTest.java
@@ -3,7 +3,7 @@ package sk.leziemevpezinku.spring.api;
 import org.junit.jupiter.api.Test;
 import sk.leziemevpezinku.spring.annotation.UnitTest;
 import sk.leziemevpezinku.spring.api.dto.EnumerationItemDTO;
-import sk.leziemevpezinku.spring.api.mapper.EnumerationItemMapper;
+import sk.leziemevpezinku.spring.mapper.EnumerationItemMapper;
 import sk.leziemevpezinku.spring.model.EnumerationItem;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
