+++ src/test/java/sk/leziemevpezinku/spring/api/EventMapperTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
 import sk.leziemevpezinku.spring.annotation.UnitTest;
 import sk.leziemevpezinku.spring.api.dto.EventDTO;
-import sk.leziemevpezinku.spring.api.mapper.EventConditionMapper;
-import sk.leziemevpezinku.spring.api.mapper.EventMapper;
+import sk.leziemevpezinku.spring.mapper.EventConditionMapper;
+import sk.leziemevpezinku.spring.mapper.EventMapper;
 import sk.leziemevpezinku.spring.model.Event;
 import sk.leziemevpezinku.spring.model.EventCondition;
 import sk.leziemevpezinku.spring.model.EventTerm;
